deleteCustomer working
